cloudfunctions/azure-functions/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
click==8.1.7
fastapi==0.114.2
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
pytest==8.3.3
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6